requirements.txt
# Core Frameworks & Libraries
streamlit
langchain
langchain-core
langchain-community
pydantic

# LLM and Embeddings (Google Gemini)
google-generativeai
langchain-google-genai

# Vector Store
faiss-cpu
python-docx
# PDF Processing
pypdf2
# OR (Choose ONE PDF library)
pymupdf
Pillow
# Utilities
python-dotenv
tiktoken
pytesseract
# Optional (Install if/when needed)
regex
tenacity
pytesseract
opencv-python-headless 
xlsxwriter # Corrected typo, no version specified initially
langdetect